+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed "Clear" button on FT searches.
 
-## [3.62.0] - 2020-06-22
+## [3.62.0] - 2020-06-22 [YANKED]
 ### Changed
 - "Clear" button on mobile filters to uncheck every brand and specification filter.